requirements-test.txt
pytest
pytest-aiohttp
pytest-asyncio==0.15.1
pytest-cov
coveralls
aioresponses
black==21.5b1
flake8==3.9.2
flake8-docstrings==1.6.0
flake8-noqa==1.1.0
isort==5.8.0
pydocstyle==6.0.0
mypy
pylint
tox
